requirements.txt
streamlit>=1.28,<2
pandas>=2.0,<3
numpy>=1.26,<2
gspread>=6.0.0
google-auth>=2.28.0
